client/requirements.txt
requests
python-dotenv
opencv-python
tqdm
pillow
